% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWAS_compileR.R
\name{FWAS_compileR}
\alias{FWAS_compileR}
\title{AWWA Free Water Audit Software (FWAS) Report Compiler}
\usage{
FWAS_compileR(path_reports, path_save = NULL, filename = NULL)
}
\arguments{
\item{path_reports}{A directory where the AWWA FWAS reports are located. This must be formatted in standard R syntax using forward slashes (/) instead of backslashes (\)}

\item{path_save}{Optional. The directory where you wish to save the compiled data as a MS Excel workbook.}

\item{filename}{Optional. The name of the file should you wish to save one.}
}
\value{
This function returns a data frame containing the summary data extracted from any version of the AWWA Free Water Audit Software (i.e. 6.1/6.0, 5.0, 4.2/4.1).
Optionally, the function can export the generated data frame as a Microsoft Excel file by specifying a directory using the "path_save" argument.
Unless specified using the "filename" argument, the saved Microsoft Excel file will be named with using the following convention "AWWA_FWAS_compiled_(timestamp).xslx".
}
\description{
AWWA Free Water Audit Software (FWAS) Report Compiler
}
\examples{
# Reference the directory of example water audits in this R package
example_audits <- system.file("example_audits_xlsx", package = "awwafwas")

# Run that list of audits through the compiler function
# Leaving the "path_save" as NULL, the function does not save the output file.
# Specifying the "path_save" as a directory will save the output file at that location (MS Excel).L
example_audits_compiled <- FWAS_compileR(path_reports = example_audits,
                                         path_save    = NULL,
                                         filename     = NULL)


}
